Project_8 QR code encoder decoder/requirements.txt
streamlit
qrcode
Pillow
opencv-python-headless
numpy
